# api/handler/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api-geteway for user and story.",
        "title": "Api Get Eway",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/destinations": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all Travel Destinations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Destinations"
                ],
                "summary": "List Travel Destinations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/destination.ListDetinationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    }
                }
            }
        },
        "/api/v1/destinations/trending": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get trending Travel Destinations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Destinations"
                ],
                "summary": "Get Trend Travel Destinations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Top Destination",
                        "name": "Top",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/destination.GetTrendDestinationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    }
                }
            }
        },
        "/api/v1/destinations/{destinationId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a single Travel Destination by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Destinations"
                ],
                "summary": "Get Travel Destination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination ID",
                        "name": "destinationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/destination.GetDestinationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    }
                }
            }
        },
...
        "/api/v1/messages": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all messages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "List Messages",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/communication.ListMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Send a message to a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Send Message to User",
                "parameters": [
                    {
                        "description": "Message",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/communication.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/communication.SendMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    }
                }
            }
        },
...
        "/api/v1/travel-tips": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add new travel tips",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TravelTips"
                ],
                "summary": "Add Travel Tips",
                "parameters": [
                    {
                        "description": "Travel Tips",
                        "name": "tips",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/communication.AddTravelTipsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/communication.AddTravelTipsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    }
                }
            }
        },
        "/api/v1/travel-tips/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get travel tips by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TravelTips"
                ],
                "summary": "Get Travel Tips",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category",
                        "name": "category",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/communication.GetTravelTipsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{userId}/statics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user statistics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/communication.GetUserStaticsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Errors"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "communication.AddTravelTipsRequest": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "communication.AddTravelTipsResponse": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "communication.Author": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "communication.GetTravelTipsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "tips": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/communication.Tip"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "communication.GetUserStaticsResponse": {
            "type": "object",
            "properties": {
                "most_popular_itinerary": {
                    "$ref": "#/definitions/communication.MostPopularItinerary"
                },
                "most_popular_story": {
                    "$ref": "#/definitions/communication.MostPopularStory"
                },
                "total_comments_received": {
                    "type": "integer"
                },
                "total_countries_visited": {
                    "type": "integer"
                },
                "total_itineraries": {
                    "type": "integer"
                },
                "total_likes_received": {
                    "type": "integer"
                },
                "total_stories": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "communication.ListMessageResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "message": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/communication.Message"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "communication.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "recipient": {
                    "$ref": "#/definitions/communication.Recipient"
                },
                "sender": {
                    "$ref": "#/definitions/communication.Sender"
                }
            }
        },
        "communication.MostPopularItinerary": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "likes_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "communication.MostPopularStory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "likes_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "communication.Recipient": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "communication.SendMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "recipient_id": {
                    "type": "string"
                },
                "sende_id": {
                    "type": "string"
                }
            }
        },
        "communication.SendMessageResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "recipient_id": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "string"
                }
            }
        },
        "communication.Sender": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "communication.Tip": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/communication.Author"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "destination.Destination": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "popular_activities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "destination.GetDestinationResponse": {
            "type": "object",
            "properties": {
                "average_cost_per_day": {
                    "type": "string"
                },
                "best_time_to_visit": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "popular_activities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "top_attractions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "destination.GetTrendDestinationResponse": {
            "type": "object",
            "properties": {
                "destinations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/destination.TrendDestination"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "destination.ListDetinationResponse": {
            "type": "object",
            "properties": {
                "destinations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/destination.Destination"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "destination.TrendDestination": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "popularity_score": {
                    "type": "string"
                },
                "trending_reason": {
                    "type": "string"
                }
            }
        },
...
        "models.Errors": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}